+++ IRaDataHolder.cs
@@ -1,18 +1,28 @@
 ﻿namespace RaDataHolder
 {
-	public interface IRaDataHolder
+	public interface IRaDataHolder : IRaDataSetResolver, IRaDataClearResolver
 	{
 		bool HasData
 		{
 			get;
 		}
 
-		void SetData(object data);
-		void ClearData();
+		IRaDataSetResolver SetRawData(object data);
+		IRaDataClearResolver ClearData();
 	}
 
 	public interface IRaDataHolder<TData> : IRaDataHolder
 	{
-		void SetData(TData data);
+		IRaDataSetResolver SetData(TData data);
+	}
+
+	public interface IRaDataSetResolver
+	{
+		IRaDataSetResolver Resolve();
+	}
+
+	public interface IRaDataClearResolver
+	{
+		IRaDataClearResolver Resolve();
 	}
 }